feature: add unit testing foe CreateItemUseCase. chore: change value type in CreateItemAttributeCommand to int. Include attribute id  in exception message when attributeId not found

+++ src/main/java/com/lps/vitalMagic/inventory/application/command/CreateItemAttributeCommand.java
@@ -2,6 +2,6 @@ package com.lps.vitalMagic.inventory.application.command;
 
 public record CreateItemAttributeCommand(
         Long attributeId,
-        Byte value
+        int value
 
 ) {}
